Add an option to mount paths provided in flags
When providing filepath on the commandline like `$ wget -O
../google.html https://www.google.com` the destination path is not bound
in the current and hence the user does not see the output file.

With the proposed implementation, the package provider should provide
flags that should contain path to files. Then, whalebrew will parse
those flags and extract volume names from it, instructing to bind those
inside the container.

There is no message added at the install time. Rationale is that the
user provides the paths himself, and hence he expects the path to be
bound. Unlike volumes that are bundled inside the image package.

I am foreseeing few issues to be addressed:
- in case a path is a file it might happen that the output is actually
  not flushed inside the host. It has been tested on OSX and seems to
  work though
- in case the path is a file and does not exist, we would expect it to
  be replicated with same user and permissions as in the container.
  Instead, an error will be raised that the file does not exist
- in case the path is ~/<something> or $HOME/<something>, it would
  probably be expected to map it to HOME in the container. Though the
  detection is quite hard to do. The HOME in the container is unknown
  and there is no trivial decision wether the intent is to handle it as
  relative to the HOME directory or if it is important that the actual
  fully provided path is forwarded

This change actually updates the in-memory package update at runtime,
which might introduce the need to introduce a runtime object to run the
command

Fixes: #22